backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
